# Overriden daysBetween implementation for ImmutableHolidayCalendar performance