Noticed when calculating lots of PVs that there was a major hot spot calculating year fractions using Business252DayCount.
Other day counts calls have a relatively fixed cost - they don't need to check individual holidays.
Whereas this uses the holiday calendar `daysBetween` which streams over every date and checks whether it is a holiday, creating lots of intermediary objects and lookups.
We can use the bit-packed lookup in ImmutableHolidayCalendar and the JDK method `Integer.bitCount()` to count business days per month in a single call.

Using some very unscientific benchmarking of the `test_yearFraction` method on `Business252DayCountTest` for 300K year fraction calls between 1 and 1000 days in the future the total time went from 51.49 seconds locally to 256ms, or just over 200 times faster.

Limiting to just the next day (as in requesting the yearFraction between today and tomorrow) 100 million times the speed up was from 27 seconds to 8s, or roughly 3 times faster